Extracting lateral methods to lateral.py
