+++ opendbc/car/mazda/carcontroller.py
@@ -1,5 +1,6 @@
 from opendbc.can import CANPacker
-from opendbc.car import Bus, apply_driver_steer_torque_limits, structs
+from opendbc.car import Bus, structs
+from opendbc.car.lateral import apply_driver_steer_torque_limits
 from opendbc.car.interfaces import CarControllerBase
 from opendbc.car.mazda import mazdacan
 from opendbc.car.mazda.values import CarControllerParams, Buttons
